Extracted components from LinkItem
