+++ frontend-new/components/InfoBox.tsx
@@ -1,0 +1,31 @@
+import { Box, Typography } from '@mui/material'
+import { BoxProps } from '@mui/system'
+import { FC, PropsWithChildren } from 'react'
+
+interface Props extends PropsWithChildren {
+  sx?: BoxProps['sx']
+  bold?: boolean
+}
+
+export const InfoBox: FC<Props> = ({ children, sx, bold = false }) => (
+  <Box
+    sx={{
+      backgroundColor: '#fff',
+      borderRadius: '32px',
+      display: 'flex',
+      alignItems: 'center',
+      px: 1,
+      height: 24,
+      mr: 1,
+      cursor: 'default',
+      '@media (min-width: 840px)': {
+        px: 2,
+        mr: 3,
+        height: 32,
+      },
+      ...sx,
+    }}
+  >
+    <Typography variant={bold ? 'h3' : 'body1'}>{children}</Typography>
+  </Box>
+)
