+++ frontend-new/components/LinkItem.tsx
@@ -1,22 +1,15 @@
 import styled from '@emotion/styled'
-import { Box, Button, IconButton, TextField, Tooltip, Typography } from '@mui/material'
-import { BoxProps } from '@mui/system'
-import {
-  ChangeEvent,
-  FC,
-  FormEvent,
-  PropsWithChildren,
-  useCallback,
-  useState,
-  useMemo,
-} from 'react'
+import { Box, IconButton } from '@mui/material'
+import { FC } from 'react'
 
 import { DeleteModal } from 'app/components/DeleteModal'
 import { TransferModal } from 'app/components/TransferModal'
-import { useUpdateLink, useDeleteLink, useModal, useClipboard } from 'app/hooks'
-import { Copy, Edit } from 'app/icons'
+import { useModal, useClipboard, useFullShortPath } from 'app/hooks'
+import { Copy } from 'app/icons'
 import { Link } from 'app/types'
 
+import { EditableDestination } from './EditableDestination'
+import { InfoBox } from './InfoBox'
 import { LinkActions } from './LinkActions'
 
 const LabelRow = styled.div`
@@ -25,75 +18,19 @@ const LabelRow = styled.div`
   align-items: center;
 `
 
-const Form = styled.form`
-  display: flex;
-  align-items: center;
-`
-
 interface Props {
   link: Link
   canEdit?: boolean
 }
 
-const InfoBox: FC<PropsWithChildren & { sx?: BoxProps['sx']; bold?: boolean }> = ({
-  children,
-  sx,
-  bold = false,
-}) => (
-  <Box
-    sx={{
-      backgroundColor: '#fff',
-      borderRadius: '32px',
-      display: 'flex',
-      alignItems: 'center',
-      px: 1,
-      height: 24,
-      mr: 1,
-      cursor: 'default',
-      '@media (min-width: 840px)': {
-        px: 2,
-        mr: 3,
-        height: 32,
-      },
-      ...sx,
-    }}
-  >
-    <Typography variant={bold ? 'h3' : 'body1'}>{children}</Typography>
-  </Box>
-)
-
 export const LinkItem: FC<Props> = ({ link, canEdit = false }) => {
-  const { id, shortpath, destination_url, owner, namespace, visits_count } = link
-  const [destination, setDestination] = useState(destination_url)
-  const [editable, setEditable] = useState(false)
+  const { id, destination_url, owner, visits_count } = link
+  const fullShortPath = useFullShortPath(link)
 
   const [transferModal, openTransferModal, closeTransferModal] = useModal()
   const [deleteModal, openDeleteModal, closeDeleteModal] = useModal()
 
-  const fullShortPath = useMemo(
-    () => `${namespace || window._trotto.defaultNamespace}/${shortpath}`,
-    [namespace, shortpath],
-  )
-  const updateLink = useUpdateLink()
-  const deleteLink = useDeleteLink()
-
-  const handleDestinationChange = useCallback(
-    (e: ChangeEvent<HTMLInputElement>) => setDestination(e.target.value),
-    [setDestination],
-  )
-
   const handleCopy = useClipboard(fullShortPath)
-
-  const handleEdit = useCallback(() => setEditable((editable) => !editable), [setEditable])
-
-  const handleSave = useCallback(
-    (e: FormEvent<HTMLFormElement>) => {
-      e.preventDefault()
-      updateLink(id, { destination })
-      handleEdit()
-    },
-    [destination, id, handleEdit, updateLink],
-  )
 
   return (
     <>
@@ -104,7 +41,6 @@ export const LinkItem: FC<Props> = ({ link, canEdit = false }) => {
           gap: 1,
           borderBottom: '1px solid #f0f0f0',
           backgroundColor: '#f6f8fa',
-          // padding: '20px 8px',
           px: 1,
           py: 2.5,
 
@@ -151,78 +87,12 @@ export const LinkItem: FC<Props> = ({ link, canEdit = false }) => {
             onDelete={openDeleteModal}
           />
         </LabelRow>
-        <Form onSubmit={handleSave}>
-          <Box
-            sx={{
-              display: 'flex',
-              flexDirection: 'row',
-              borderRadius: '32px',
-              background: '#fff',
-              mr: 1,
-              '@media (min-width: 840px)': {
-                mr: 3,
-              },
-              flexGrow: 1,
-            }}
-          >
-            <TextField
-              id='destination'
-              placeholder='Keyword'
-              value={destination}
-              onChange={handleDestinationChange}
-              disabled={!editable}
-              sx={{
-                height: 24,
-                flexGrow: 1,
-                '@media (min-width: 840px)': {
-                  height: 32,
-                },
-              }}
-            />
-            {editable && (
-              <Button
-                className='button'
-                variant='contained'
-                type='submit'
-                sx={{
-                  backgroundColor: '#000',
-                  height: 24,
-                  '@media (min-width: 840px)': {
-                    height: 32,
-                  },
-                }}
-              >
-                Save
-              </Button>
-            )}
-          </Box>
-          <Tooltip title={!canEdit && 'You don’t have permission to modify this go link'}>
-            <span>
-              <IconButton
-                onClick={handleEdit}
-                sx={{
-                  opacity: editable ? 1 : 0.25,
-                  '&:disabled': {
-                    opacity: 0.1,
-                  },
-                }}
-                disabled={!canEdit}
-              >
-                <Edit />
-              </IconButton>
-            </span>
-          </Tooltip>
-        </Form>
+        <EditableDestination id={id} destinationUrl={destination_url} disabled={!canEdit} />
       </Box>
       {transferModal && (
         <TransferModal open={transferModal} onClose={closeTransferModal} link={link} />
       )}
-      <DeleteModal
-        open={deleteModal}
-        onClose={closeDeleteModal}
-        onDelete={deleteLink}
-        link={link}
-      />
+      <DeleteModal open={deleteModal} onClose={closeDeleteModal} link={link} />
     </>
   )
 }
