+++ frontend-new/types/index.ts
@@ -3,6 +3,6 @@ export * from './user'
 
 declare global {
   interface Window {
-    _trotto: { [name: string]: unknown }
+    _trotto: { defaultNamespace: string; [name: string]: unknown }
   }
 }
